Initial prototype: display a recipe
Added code which fetches complete information for a small number of recipes. The recipe, the output item, the ingredients, and buy/sell information for all items. Added code to calculate the profit (if item is tradeable) or cost (if not) of crafting the item. Displayed the recipe, output item, and value in a list view.